fix: attempt optimize EXISTS subqueries
MySQL8 engine seems to struggle with an EXISTS subquery that is used in
most views of redmine.

https://dev.mysql.com/doc/refman/8.4/en/subquery-optimization-with-exists.html
Seems to indicate we can help the optimizer by including the outer
condition in the inner query.

In particular this quote:
> A very useful optimization is to “inform” the subquery that the only
> rows of interest are those where the inner expression inner_expr is
> equal to outer_expr. This is done by pushing down an appropriate
> equality into the subquery's WHERE clause to make it more restrictive.

This would be an example diff of the produced SQL:
```diff
SELECT COUNT(*)
FROM `issues`
INNER JOIN `projects` ON `projects`.`id` = `issues`.`project_id`
INNER JOIN `issue_statuses` ON `issue_statuses`.`id` = `issues`.`status_id`
WHERE (
    projects.status <> 9
    AND EXISTS (
        SELECT 1 AS one
        FROM enabled_modules em
        WHERE em.project_id = projects.id
        AND em.name='issue_tracking'
+       projects.status <> 9
    )
)
AND ((
    issues.status_id IN (
        SELECT id FROM issue_statuses WHERE is_closed=FALSE
    ))
)
```